pyproject.toml
[tool.poetry]
name = "decipher"
version = "0.1.11"
description = "Utilities for Decipher"
authors = ["Thorvald Molthe Ballestad <thorvald@simula.no>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pandas = "^1.5.0"
numpy = "^1.24.0"
scikit-learn = "^1.2.0"
pandas-stubs = "^1.5.3.0"
PyYAML = "^6.0"
types-PyYAML = "^6.0.12.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
pylint = "^2.17.1"
mypy = "^1.1.1"
pytest = "^7.2.2"


[tool.poetry.group.sparse.dependencies]
scipy = "^1.10.1"

[tool.isort]
profile = "black"

[[tool.mypy.overrides]]
module = [
    "sklearn.*",
    "scipy.sparse",
]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
